requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.1
django-appconf==1.0.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-imagekit==4.0.2
djangorestframework==3.11.1
filelock==3.0.12
httpie==2.2.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
lazy-object-proxy==1.4.3
pilkit==2.0
Pillow==7.2.0
pkg-resources==0.0.0
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.31
wrapt==1.12.1
zipp==3.2.0
